syscall and isr research
Made some changes for maing system calls and handling isr reqs.

+++ sys/mm/pgtable.c
@@ -196,7 +196,7 @@ uint64_t insert_page(pml4_t *cr3, uint64_t virtual_address, uint64_t permissions
     // Get the page table using this pml4 and virtual address
     pt_t* page_table = (pt_t*) PHYS_TO_VIRT(get_pt(cr3, virtual_address));
     // Get a new page to insert into the table
-    uint64_t page = kmalloc_pg() | permissions;
+    uint64_t page = (uint64_t)kmalloc_pg() | permissions;
     // Get the page table offset from the virtual address 
     uint64_t pt_index = extract_table(virtual_address);
     page_table->entries[pt_index] = page;
